+++ tests/test_API_TC_CS.v
@@ -109,3 +109,11 @@ Check (fun x : C1 => x 3).
 
 Elpi Query lp:{{coq.coercion.db L}}.
 
+Axiom C3 : nat -> Type.
+Axiom nuc : forall x, C1 -> C3 x.
+
+Set Warnings "+uniform-inheritance".
+Elpi Query lp:{{ @nonuniform! => @reversible! => coq.coercion.declare (coercion {coq.locate "nuc"} _ _ _) }}.
+
+About nuc.
+